# feat: only retrieve the "key" field when needed