--- a/pg_search/src/index/mod.rs
+++ b/pg_search/src/index/mod.rs
@@ -15,7 +15,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-pub mod score;
 pub mod search;
 pub mod state;