… pg_search/src/index/score.rs
… pg_search/tests/bm25_search.rs
… pg_search/tests/documentation.rs