## What

Be more intelligent about when we decode the "key" field from Tantivy's fast_field index.  

We only need the key field when 
   
   1) the user expects it to be returned (via the `score_bm25()` function), and 
   2) to do tie-breaks between docs with the same score. 

Regarding (2), if the key field is a string and we're not returning it (but we do need to tie-break), then we elide reconstructing the whole key string and instead use the term ordinal as the tie-breaker. 

Below are some performance numbers, on my workstation, using a dataset of around 51M documents.  It's worth noting that the current `dev` branch (labeled "before") doesn't actually honor the `stable_sort` flag.  This PR does, but only in the case where it "makes sense", which is the `.search(), no limit` case.


These numbers use queries such as:

```sql
SELECT count(*) FROM idxname.search('body:(beer wine cheese)');
SELECT count(*) FROM idxname.score_bm25('body:(beer wine cheese)');
```

Without a limit, they return 179,871 rows.

| style                               | before      | after       |
|-------------------------------------|-------------|-------------|
| .search(), no limit, ss=false       | 1377.903 ms | 498.527 ms* |
| .search(), limit=1000, ss=false     | 869.195 ms  | 13.316 ms*  |
| .score_bm25(), no limit, ss=false   | 1329.254 ms | 1301.576 ms |
| .score_bm25(), limit=1000, ss=false | 839.794 ms  | 22.327 ms*  |
| .search(), no limit, ss=true        | 1372.652 ms | 489.407 ms* |

We can see this PR improves all the cases, quite drastically, except for the `score_bm25(), no limit` case.  That cases shows a little improvement to do the elimination of a `.collect::<Vec<_>>()`, but is otherwise still functionally equivalent to the `dev` branch -- that's why it's not greatly improved.

## Why

To improve search performance.

## How

By asking Tantivy to do the minimal amount of work to answer a query based on its `SearchConfig`urations and also based on the call points.  Internally, there were some API changes to `SearchState`.

There's further work that can happen to eliminate a few `.collect::<Vec<_>>()` that collect up the results of a tantivy search, but those can happen in a subsequent PR after this is merged.

## Tests

I've had to adjust a few tests that use `idxname.search()` since the results are no longer stable.  In some cases I added `stable_sort => true` to the query, and in others I adjusted the test code itself to account for unpredictable ordering.